modify script to the files are names after the number of contributing loci

+++ scripts/build_population_for_sim.py
@@ -10,7 +10,7 @@ beta = int(snakemake.params['beta'])
 
 polygenicity_params_file = snakemake.input['polygenicity_params'] 
 
-pi =  snakemake.params['pi'].astype(int)
+pi =  int(snakemake.params['pi'])
 #pi = pd.read_csv(polygenicity_params_file,header=None, usecols=[int(pi_option)]).values[0][0]
 
 #get the actual values
